docker/requirements.txt
    fastapi
    uvicorn[standard]
    python-multipart
    chatterbox-tts
    torchaudio
    torch